Crate euler

Euler

This crate is part of xfbs/euler, which aims to solve Project Euler problems in a variety of languages.

The functions and traits of this crate are sharedbetween solutions solutions and implement generic, reusable algorithms. Tests should be available for all functions.

If you clone the repository, and navigate to this library (which is in /lib/rust/euler), there is a Makefile that supports the following targets:

make test to build and run all the tests of this library.

make fmt to run cargo fmt, the rust code formatter, over all of the code in this library.

make doc to generates these docs. The output will be placed in /doc/lib/rust in the repository.

All code in this library has been developed and tested with Rust 1.22.1. Since rust is still under development, things may break with future releases of the compiler.

Structs

Digits
Iterator over the digits of a number.

Permutations
An iterator over the permutation of elements.

Prime
Prime number generator.

Rotations
Iterator over rotations of a number.

Sieve
Prime sieve.

Traits

Gcd
The greatest common divisor of two types.

IsPalindrome
Trait to check if a number is a palindrome.

Lcm
The least common multiple of two types.

ToDigits
Trait to convert a number into an iterator over its digits.

ToDivisors
Trait to turn a number into an Iterator over its divisors.

ToPalindrome
Trait to turn a number into a palindrome.

ToRotations
Trait to generate an iterator over the rotations of a number.

Functions

factorial
Returns the nth factorial.

permutations
Returns an iterator over the nth permutation of count elements.