Struct euler::Prime

pub struct Prime { /* fields omitted */ }

Prime number generator.

Methods

impl Prime

pub fn new() -> Self

Construct a new prime number generator.

Examples

use euler::Prime;

let mut primes = Prime::new();
assert_eq!(primes.nth(4), 7);

pub fn nth(&mut self, n: usize) -> u64

Generates nth prime.

Examples

use euler::Prime;

let mut primes = Prime::new();
assert_eq!(primes.nth(8), 19);
assert_eq!(primes.nth(9), 23);
assert_eq!(primes.nth(10), 29);

pub fn check(&mut self, num: u64) -> bool

Checks if a number is prime.

Tries to divide by all primes smaller or equal to the square root of the number, returns true if it found a divisor or false if the number is prime.

Examples

use euler::Prime;

let mut primes = Prime::new();
assert_eq!(primes.check(97), true);
assert_eq!(primes.check(99), false);