Struct euler::Sieve

pub struct Sieve { /* fields omitted */ }

Prime sieve.

Implements IntoIter so that it can be used as an iterator over the primes.

Examples

use euler::Sieve;

let sieve = Sieve::new(100);
assert_eq!(sieve.into_iter().count(), 25);

Methods

impl Sieve

pub fn new(max: usize) -> Self

Create a new prime sieve, and sieve all numbers up to max.

Sieves all numbers from 0 until max and returns a Sieve instance with the resulting matrix.

Examples

use euler::Sieve;

let sieve = Sieve::new(1000);
assert_eq!(sieve.check(17), Some(true));
assert_eq!(sieve.check(20), Some(false));
assert_eq!(sieve.check(1001), None);

pub fn check(&self, num: u64) -> Option<bool>

Checks if a given number is prime or not.

Examples

use euler::Sieve;

let sieve = Sieve::new(1000);
assert_eq!(sieve.check(17), Some(true));
assert_eq!(sieve.check(20), Some(false));
assert_eq!(sieve.check(1001), None);