Crate euler

Euler

This crate is part of xfbs/euler, which aims to solve Project Euler problems in a variety of languages.

The functions and traits of this crate are sharedbetween solutions solutions and implement generic, reusable algorithms. Tests should be available for all functions.

If you clone the repository, and navigate to this library (which is in /lib/rust/euler), there is a Makefile that supports the following targets:

make test

Builds and runs all the tests of this library.

make fmt

Runs the crystal code formatter over all of the code in this library.

make doc

Generates these docs, the output will be placed in doc/lib/crystal in the repository.

All code in this library has been developed and tested with Rust 1.22.1. Since rust is still under development, things may break with future releases of the compiler.

Structs

Prime

Sieve

Traits

Gcd
The greatest common divisor of two types.

IsPalindrome

Lcm
The least common multiple of two types.

ToDigits

ToDivisors

ToPalindrome

ToRotations

Functions

factorial
Returns the nth factorial.

permutations